+++ src/main/java/inu/spp/cryptocoffee/auth/user/service/UserService.java
@@ -3,8 +3,8 @@ package inu.spp.cryptocoffee.auth.user.service;
 import inu.spp.cryptocoffee.auth.user.entity.UserEntity;
 import inu.spp.cryptocoffee.auth.user.jwt.JwtUtil;
 import inu.spp.cryptocoffee.auth.user.repository.UserRepository;
-import inu.spp.cryptocoffee.auth.user.controller.UserAuthorityRequestDto;
-import inu.spp.cryptocoffee.auth.user.controller.UserPageResponseDto;
+import inu.spp.cryptocoffee.auth.user.dto.UserAuthorityRequestDto;
+import inu.spp.cryptocoffee.auth.user.dto.UserPageResponseDto;
 import inu.spp.cryptocoffee.domain.company.entity.CompanyEntity;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
